% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft.R
\name{av_audio_fft}
\alias{av_audio_fft}
\title{Calculate FFT}
\usage{
av_audio_fft(audio)
}
\arguments{
\item{audio}{audio or video input file with sound for the output video}
}
\description{
Convert audio into frequency domain.
}
