% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{av_video_info}
\alias{av_video_info}
\title{Video Info}
\usage{
av_video_info(file)
}
\arguments{
\item{file}{path to an existing file}
}
\description{
Get video info
}
\seealso{
Other av: \code{\link{av_capture_graphics}},
  \code{\link{av_encode_video}}, \code{\link{av_encoders}}
}
\concept{av}
