% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture.R
\name{capturing}
\alias{capturing}
\alias{av_capture_graphics}
\title{Record Video from Graphics Device}
\usage{
av_capture_graphics(expr, output = "output.mp4", width = 720,
  height = 480, framerate = 1, filter = "null", verbose = TRUE,
  ...)
}
\arguments{
\item{expr}{an R expression that generates the graphics to capture}

\item{width}{width in pixels of the graphics device}

\item{height}{height in pixels of the graphics device}

\item{...}{extra graphics parameters passed to \code{\link[=png]{png()}}}
}
\description{
Runs the expression and captures all plots into a video.
}
\examples{
\donttest{
library(gapminder)
library(ggplot2)
makeplot <- function(){
  datalist <- split(gapminder, gapminder$year)
  lapply(datalist, function(data){
    p <- ggplot(data, aes(gdpPercap, lifeExp, size = pop, color = continent)) +
      scale_size("population", limits = range(gapminder$pop)) + geom_point() + ylim(20, 90) +
      scale_x_log10(limits = range(gapminder$gdpPercap)) + ggtitle(data$year) + theme_classic()
    print(p)
  })
}

# Play 1 plot per sec, and use an interpolation filter to convert into 10 fps
video_file <- file.path(tempdir(), 'output.mp4')
av_capture_graphics(makeplot(), video_file, 1280, 720, res = 144, filter = 'framerate=fps=10')
av::av_video_info(video_file)
utils::browseURL(video_file)}
}
\seealso{
Other av: \code{\link{encoding}}, \code{\link{formats}},
  \code{\link{info}}, \code{\link{logging}}
}
\concept{av}
