% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{demo}
\alias{demo}
\alias{av_demo}
\title{Demo Video}
\usage{
av_demo(width = 960, height = 720, framerate = 5, verbose = TRUE,
  ...)
}
\arguments{
\item{width}{width in pixels of the graphics device}

\item{height}{height in pixels of the graphics device}

\item{framerate}{video framerate in frames per seconds. This is the input fps, the
output fps may be different if you specify a filter that modifies speed or interpolates
frames.}

\item{verbose}{emit some output and a progress meter counting processed images. Must
be \code{TRUE} or \code{FALSE} or an integer with a valid \href{av_log_level}{log_level}.}

\item{...}{other parameters passed to \link{av_capture_graphics}.}
}
\description{
Generates random video for testing purposes.
}
\seealso{
Other av: \code{\link{capturing}}, \code{\link{encoding}},
  \code{\link{formats}}, \code{\link{info}},
  \code{\link{logging}}
}
\concept{av}
