% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\name{av_encoders}
\alias{av_encoders}
\alias{av_decoders}
\alias{av_filters}
\alias{av_muxers}
\alias{av_demuxers}
\title{AV Formats}
\usage{
av_encoders()

av_decoders()

av_filters()

av_muxers()

av_demuxers()
}
\description{
List supported filters, codecs and container formats.
}
\details{
Encoders and decoders convert between raw video/audio frames and compressed stream
data for storage or transfer. However such a compressed data stream by itself does
not constitute a valid video format yet. Muxers are needed to interleave one or more
audio/video/subtitle streams, along with timestamps, metadata, etc, into a proper
file format, such as mp4 or mkv.

Conversely, demuxers are needed to read a file format into the seperate audio/video
streams, for subsequent decoding. Most operating systems include native support for
demuxing and decoding commonly used file formats and codecs. However for encoding
and muxing, ffmpeg must have been configured with the proper external libraries.
}
\seealso{
Other av: \code{\link{av_encode_video}},
  \code{\link{av_video_info}}
}
\concept{av}
