% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{info}
\alias{info}
\alias{av_video_info}
\title{Video Info}
\usage{
av_video_info(file)
}
\arguments{
\item{file}{path to an existing file}
}
\description{
Get video info such as width, height, format, duration and framerate.
This may also be used for audio input files.
}
\seealso{
Other av: 
\code{\link{capturing}},
\code{\link{demo}()},
\code{\link{encoding}},
\code{\link{formats}},
\code{\link{logging}}
}
\concept{av}
