% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{logging}
\alias{logging}
\alias{av_log_level}
\title{Logging}
\usage{
av_log_level(set = NULL)
}
\arguments{
\item{set}{new \href{https://www.ffmpeg.org/doxygen/4.0/group__lavu__log__constants.html}{log level} value}
}
\description{
Get or set the \href{https://www.ffmpeg.org/doxygen/4.0/group__lavu__log__constants.html}{log level}.
}
\seealso{
Other av: 
\code{\link{capturing}},
\code{\link{demo}()},
\code{\link{encoding}},
\code{\link{formats}},
\code{\link{info}}
}
\concept{av}
