% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{av_convert}
\alias{av_convert}
\title{Convert to Video}
\usage{
av_convert(input, output)
}
\arguments{
\item{input}{vector with image or video files}

\item{output}{filename of the output file. Container type is inferred from
the file extension which should typically be \code{.mkv} or \code{.mp4}.}
}
\description{
Convert Images into Video
}
