% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/video.R
\name{av_encode_video}
\alias{av_encode_video}
\title{Create Video}
\usage{
av_encode_video(input, output = "video.mp4", framerate = 1,
  filter = "null", codec = NULL, verbose = TRUE)
}
\arguments{
\item{input}{a vector with images such as png files.}

\item{output}{name of the output file. File extension must correspond to a known
container format such as \code{mp4}, \code{mkv}, \code{mov}, or \code{flv}.}

\item{framerate}{video framerate in frames per seconds.}

\item{filter}{a string defining an ffmpeg filter graph}

\item{codec}{name of the video codec as listed in \link{av_encoders}. Most
formats default to h264 which has excellent compression and works out of the box on
all \href{https://caniuse.com/#search=h264}{modern browsers} and operating systems.}

\item{verbose}{emit some output and a progress meter counting processed images. Must
be \code{TRUE} or \code{FALSE} or an integer with a valid \href{av_log_level}{log_level}.}
}
\description{
Encodes a series of images into a video.
}
