% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{av_log_level}
\alias{av_log_level}
\title{Logging}
\usage{
av_log_level(set = NULL)
}
\arguments{
\item{set}{new [log level](https://www.ffmpeg.org/doxygen/4.0/group__lavu__log__constants.html) value}
}
\description{
Get or set the [log level](https://www.ffmpeg.org/doxygen/4.0/group__lavu__log__constants.html).
}
