% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{av_video_images}
\alias{av_video_images}
\title{Split Video into PNG}
\usage{
av_video_images(video, destdir = tempfile(), fps = 1)
}
\arguments{
\item{video}{an input video}

\item{destdir}{directory where to save the png files}

\item{fps}{sample rate of images. Use \code{NULL} to get all images.}
}
\description{
Convert a video file in a set of PNG images.
}
